Minor cleanups to prepare HgiVulkan to build on macOS and Windows. (imageless framebuffer extension is no longer used, but forgot to remove it a few changelists ago)
(Internal change: 2119386)

+++ pxr/imaging/hgiVulkan/computePipeline.cpp
@@ -86,7 +86,7 @@ HgiVulkanComputePipeline::HgiVulkanComputePipeline(
 
     _vkDescriptorSetLayouts = HgiVulkanMakeDescriptorSetLayouts(
         device, {setInfo}, desc.debugName);
-    pipeLayCreateInfo.setLayoutCount = _vkDescriptorSetLayouts.size();
+    pipeLayCreateInfo.setLayoutCount = (uint32_t)_vkDescriptorSetLayouts.size();
     pipeLayCreateInfo.pSetLayouts = _vkDescriptorSetLayouts.data();
 
     TF_VERIFY(
